requirements.txt
opencv-python-headless==4.5.5.62
yacs==0.1.8
albumentations==1.1.0
matplotlib==3.4.2
tqdm==4.62.2
plyfile==0.7.4
scipy==1.7.1
pandas==1.3.1
numpy==1.21.2
Pillow==9.0.0
torchinfo
